requirements.txt
asyncio
python-socketio
aiohttp
uvicorn[standard]